app_desktop/requirements.txt
customtkinter
pyserial
matplotlib
